BOJ 16693 : Pizza Deal
There’s a pizza store which serves pizza in two sizes: either a pizza slice, with area A1 and price P1, or a circular pizza, with radius R1 and price P2.

You want to maximize the amount of pizza you get per dollar. Should you pick the pizza slice or the whole pizza?