백준 21633 : Bank Transfer
Tanya has an account in "Redgotts" bank. The bank has the commission to transfer money to "Bluegotts" bank that her friend Vanya has the account in.

Tanya has read her bank rules and learned the following:

The commission for the bank transfer is 25 tugriks plus 1% of the sum transferred. However, the commission is never smaller than 100 tugriks, and cannot exceed 2000 tugriks.

Tanya is planning to transfer k tugriks to Vanya. Help her find out what the commission would be.