백준 25784 : Easy-to-Solve Expressions
When one looks at a set of numbers, one usually wonders if there is a relationship among them? This task is more manageable if there are only three numbers.

Given three distinct positive integers, you are to determine how one can be computed using the other two. Print 1 if any of the three numbers is the sum of the other two numbers, print 2 if any of the three numbers is the product of the other two numbers, print 3 otherwise. Assume that exactly one of these three messages will apply.