백준 25881 : Electric Bill
To encourage customers to conserve energy (and protect the environment), the electric companies typically have a lower rate for the first 100 kilowatt-hour (KWH) of electric usage and a higher rate for the additional usage (KWH is a derived unit of energy equal to 3.6 megajoules).

Given the rate (per KWH) for the first 1000 KWH usage, the rate (per KWH) for the additional usage, and a customer's energy consumption, you are to determine the charges (bill) for the customer.