백준 25991 : Lots of Liquid
You work at a warehouse that sells chemical products, where somebody just placed an order for all the Boron Acetate Phosphoric Carbonate (BAPC) that you have in store. This liquid is stored in many separate lots, in cube-shaped containers, but your client requires the order to be delivered in a single cube-shaped container that fits all the BAPC liquid perfectly. What should be the size of this container?