백준 6825 : Body Mass Index
The Body Mass Index (BMI) is one of the calculations used by doctors to assess an adult's health. The doctor measures the patient's height (in meters) and weight (in kilograms), then calculates the BMI using the formula

BMI = weight/(height x height).

Write a program which prompts for the patient's height and weight, calculates the BMI, and displays the corresponding message from the table below.