백준 5358 : Football Team
The PLU football coach must submit to the NCAA officials the names of all players that will be competing in NCAA Division || championship game. Unfortunately his computer keyboard malfunctioned and interchanged the letters 'i' and 'e'. Your job is to write a program that will read all the names and print the names with the correct spelling.