백준 26532 : Acres
You have a rectangular field you want to plant with corn. You know the dimensions of the field in yards, and you know that one bag of corn seed will cover 5 acres. Having passed all your elementary math courses you also know that 4840 square yards is equal to 1 acre. So, given the dimensions of the rectangular field in yards you must compute the number of bags of corn seed you need to plant the entire field.