백준 26495 : Big Number
One of the professors at PLU has had a lot of trouble reading student programs. He just cannot read that small print. Your job is to write a program that will read a positive integer and rewrite the number in large block format. The block format for each digit is given below.

0000    1 2222 3333 4  4 5555 6666 7777 8888 9999
0  0    1    2    3 4  4 5    6       7 8  8 9  9
0  0    1 2222 3333 4444 5555 6666    7 8888 9999
0  0    1 2       3    4    5 6  6    7 8  8    9
0000    1 2222 3333    4 5555 6666    7 8888    9