백준 25893: Majestic 10
The movie "Magnificent 7" has become a western classic. Well, this year we have 10 coaches training the UCF programming teams and once you meet them, you'll realize why they are called the "Majestic 10"! The number 10 is actually special in many different ways. For example, in basketball, they keep track of various statistics(points scored, rebounds, etc.) and if a player has 10+ (10 or more) in a particular stat, they call it a double.

Given three stats for a basketball player, you are to determine how many doubles the player has, i.e., how many of the stats are greater than or equal to 10.